requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
django-cors-headers==4.3.1
django-allauth==0.61.1
django-filter==24.1
stripe==8.2.0
python-decouple==3.8
reportlab==4.0.9
weasyprint==60.2
drf-spectacular==0.27.1
pytest-django==4.8.0
django-import-export==3.3.7
django-ratelimit==4.1.0
gunicorn==21.2.0